Changed lots of things added many options in the slideshow and added mechanisms
Most features are added - now only need to c+p info and lines of code to add info boxes. Carcinogen quiz is still not added in the python nor javascript until we get more information from the guy we helping (if you are seeing this can you tell us how to do the carcinogen quiz? could not contact you)